Moved simple and tripod engineControl to one node
