+++ catkin_ws/src/acr_hardware/src/engineControl/normal.hpp
@@ -1,21 +1,18 @@
 #include "ros/ros.h"
 #include "geometry_msgs/Twist.h"
 
-#include "../engineControl/engineControl.hpp"
-
-#define ENGINE_TIMEOUT 1000
-#define ENGINE_LEFT ENGINE0
-#define ENGINE_RIGHT ENGINE1
-
-unsigned int lastMessage = millis();
+#include "engineControl.hpp"
 
 using namespace ros;
+
+#define ENGINE_LEFT ENGINE0
+#define ENGINE_RIGHT ENGINE1
 
 /**
  * Callback for received messages on 'sensor_modules'
  * Writes to the pins if 'SENSOR_IDLE' or 'SENSOR_INTERACT' is received
  */
-void moduleCallback(const geometry_msgs::Twist& msg) {
+void normalCallback(const geometry_msgs::Twist& msg) {
 	if (msg.linear.x > 1 || msg.linear.x < -1) {
 		ROS_WARN("Invalid linear.x, must be between -1 and 1 (now is %f)", msg.linear.x);
 		return;
@@ -50,24 +47,4 @@ void moduleCallback(const geometry_msgs::Twist& msg) {
 			runEngine(ENGINE_RIGHT, speed);
 		}
 	}
-}
-
-int main(int argc, char **argv) {
-	setupPins();
-	init(argc, argv, "simpleEngineControl");
-	NodeHandle nh;
-
-	Subscriber sub = nh.subscribe("cmd_vel", 10, moduleCallback);
-
-	Rate loop_rate(10.f);	// 10 Hz
-	
-	while(ros::ok()) {
-		if (millis() - lastMessage > ENGINE_TIMEOUT) {
-			stopEngine(ENGINE_LEFT);
-			stopEngine(ENGINE_RIGHT);
-		}
-		spinOnce();	
-		loop_rate.sleep();
-	}			
-  return 0;
 }
